Move boltz client to pkg
