+++ pkg/boltz/go.mod
@@ -1,0 +1,8 @@
+module github.com/ArkLabsHQ/fulmine/pkg/boltz
+
+go 1.24.2
+
+require (
+	github.com/gorilla/websocket v1.5.3
+	github.com/mitchellh/mapstructure v1.5.0
+)
